+++ spec/aws/lex/conversation_spec.rb
@@ -6,6 +6,80 @@ describe Aws::Lex::Conversation do
 
   subject { described_class.new(event: event, context: lambda_context) }
 
+  describe '#active_context' do
+    context 'when an active context exists' do
+      before(:each) do
+        subject.active_context!(name: 'test')
+      end
+
+      it 'returns the context' do
+        expect(subject.active_context(name: 'test')).to be_a(Aws::Lex::Conversation::Type::Context)
+      end
+    end
+
+    context 'when an active context does not exist' do
+      it 'returns nil' do
+        expect(subject.active_context(name: 'test')).to be_nil
+      end
+    end
+  end
+
+  describe '#active_context?' do
+    context 'when an active context exists' do
+      before(:each) do
+        subject.active_context!(name: 'test')
+      end
+
+      it 'returns true' do
+        expect(subject.active_context?(name: 'test')).to be(true)
+      end
+    end
+
+    context 'when an active context does not exist' do
+      it 'returns false' do
+        expect(subject.active_context?(name: 'test')).to be(false)
+      end
+    end
+  end
+
+  describe '#active_context!' do
+    context 'when an active context exists' do
+      before(:each) do
+        subject.active_context!(name: 'test')
+      end
+
+      it 'updates the existing context attributes' do
+        subject.active_context!(
+          name: 'test',
+          turns: 2,
+          seconds: 10,
+          attributes: {
+            foo: 'bar'
+          }
+        )
+        instance = subject.active_context(name: 'test')
+        expect(instance.time_to_live.time_to_live_in_seconds).to eq(10)
+        expect(instance.time_to_live.turns_to_live).to eq(2)
+        expect(instance.context_attributes).to eq(foo: 'bar')
+      end
+    end
+
+    context 'when an active context does not exist' do
+      it 'creates a new active context and returns it' do
+        expect(subject.active_context?(name: 'test')).to be(false)
+        subject.active_context!(
+          name: 'test',
+          turns: 2,
+          seconds: 10,
+          attributes: {
+            foo: 'bar'
+          }
+        )
+        expect(subject.active_context?(name: 'test')).to be(true)
+      end
+    end
+  end
+
   describe '#checkpoint!' do
     context 'when an existing checkpoint does not exist' do
       before(:each) do